+++ src/prepare.js
@@ -2,13 +2,15 @@
  * Prepares a response body
  *
  * @method prepare
- * @param  {Mixed}  data   [Optional] Response body "data"
+ * @param  {Mixed}  arg    [Optional] Response body "data"
  * @param  {Object} error  [Optional] Error instance
  * @param  {Number} status HTTP status code
  * @return {Object}        Standardized response body
  */
-function prepare ( data, error, status ) {
-	if ( data !== null ) {
+function prepare ( arg, error, status ) {
+	var data = clone( arg, true );
+
+	if ( arg !== null ) {
 		error = null;
 	}